Adds basic pickup logic and fixes incorrect setting of world bounds.
Level now has a group of pickup objects (currently FlxSprites) as a member.
These disappear when the player presses Z and overlaps one of them.

The boundaries of the level are now properly extended when adding the sprites.
FlxG.worldBounds is extended by this rectangle rather than set to it.
Previously the worldBounds were being set to FlxRect(0, 0, 0, 0), resulting in all
collisions anywhere being processed. Now this is properly limited to the level's area.