Adds a UI element showing the player's inventory.
The class is `turtleback.states.play.ui.InventoryUI`.
It provides a mechanism to show an image and a count for each type
of item the player can collect.

Though the count does update currently, the inventory data management that
will be handled by the player class has not been programmed yet.
As a temporary measure, the PlayState is keeping a count for one item type.