Adds classes for Pickup items and the player's inventory.
Pickup class just defines a type:String field on top of sprite.
It is used when picking the object up to update the correct inventory slot.

Inventory keeps track of what items the player has collected.
It also updates the inventory UI when new item types should be added
or the count of an item type should be changed.

The inventory is a public member of the player, and is interacted with
directly in that way. An inventory UI must be associated with it after
construction.