Separates logic for the player's state updating and the player changing state.
A new method is added to the player for changing their state. This performs
some one-time actions that are relevant when changing state but not during
update loops. As an example, when the player is idle they should already be
stopped, so we do not need to set their velocity to zero every update.

In order to avoid a lot of cumbersome argument passing, a property for
the current motion direction has also been added to the player.