Adds an thorough representation of a script to the cutscene state.
A script for a scene is now represented as a list of sequences of lines
delivered by a specific speaker.

There are currently two scripts, a pre-gameplay and post-gameplay script.

The game now loops fully around, from the pre-gameplay cutscene to the
gameplay to the post-gameplay scene. After the post-gameplay scene it
restarts.