Changes level data specification and moves pickup group to PlayState.
Current data specification for each state includes environment information
under the key 'level', an array of pickups under the key 'pickups',
and an array of item quotas under the key 'quotas'. The data requirements
for a pickup have also been formalized in the code.

Managing the group of pickups is not the responsibility of the PlayState.
Currently this is deemed more appropriate then the Level object being
responsible for that.