Changes how InventoryUI is organized and reduces size.
Item image sprites and count sprites are now part of a single
sprite group each, as opposed to being loosely associated via a typedef.

The inventory UI also only appears to be as large as it needs now,
rather than being the width of the screen.
In addition, the inventory UI is specifically does not move or collide.