Changes the abstraction of a Level to a series of tiles.
Previously, a level contained a sequence of simple backgrounds.
The ground was going to be defined as a member of a whole level.

Now each tile in the level will have a background and a section of the
ground corresponding to its background section. The intuitive connection
between the background art and the ground is the motivating factor for this.

Currently level tiles can support a flat ground object only.