Adds a system for handling non-level ground surfaces.
The primary component of the system is a new object class that represents
the ground itself. Each instance knows how to find the ground height at every
x position within it.

The relevant segments are found through overlap detection and logic has been
added in both the Player class and the GroundObject class for when overlap is
discovered.