Makes various FlxObject-related changes.
Sets some objects to not be solid or not move in order to avoid
potential collision checks or motion calculations in the update step
where appropriate. This is more of a declaration of intent for those
objects than a performance consideration at this point.

Additionally, the construction of FlxObjects and the act of adding them
to the group they belong to is more consistently separeated.