Moves all ground collision object into the GroundObject class.
There is now a single static member of this class that handles collisions
if they occur and returns a boolean indicating whether the collision
happened. This leaves other objects free to have a callback for if the
collision takes place.

The logic was taken from two methods in the Player class and one other
method in the GroundObject class, all of which have been removed in
favor of this.