Adds a base state for the game and a simple state machine for the states.
The base state is mainly currently used to send a signal when the state ends.
The signal is received by the `Main` class (OpenFL root sprite) to change the
state through `flixel.FlxG`.

A necessary change made alongside creating the state base class and the
state machine is that the state ID is no longer hard-coded into the play
state or the cutscene state, it is a property of the state base class.